Fixed stuck notes
The reason the keys were getting stuck down was because the talking board was trying to play sounds for the notes that were spawned, and the sound file's indexes were not specified in the array. I commented out the line because the talking board shouldn't be playing any sounds in the first place.